restructures some of the code
